+++ NEWS.md
@@ -1,3 +1,9 @@
+# welo 0.1.2
+
+* Included the possibility of downloading data from http://www.tennis-data.co.uk/ via tennis_data function
+* Introduced WElo and Elo plots via welo_plot function
+* Solved some typos
+
 # welo 0.1.1
 
 * Improved calculation speed.